+++ src/engine/platform/window.rs
@@ -1,37 +1,43 @@
+use std::sync::Arc;
+
 use winit::error::RequestError;
 use winit::event::WindowEvent;
 use winit::event_loop::ActiveEventLoop;
-use winit::window::{Window, WindowAttributes};
+use winit::window::Window;
+use winit::window::WindowAttributes;
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
 pub struct WindowManager {
-    window: Option<Box<dyn Window>>,
+    pub window: Arc<dyn Window>,
 }
 
 impl WindowManager {
-    pub fn create_window(&mut self, event_loop: &dyn ActiveEventLoop) -> Result<(), RequestError> {
-        if self.window.is_some() {
-            return Ok(());
-        }
-
+    pub fn new(event_loop: &dyn ActiveEventLoop) -> Result<Self, RequestError> {
         let attributes: WindowAttributes = WindowAttributes::default().with_title("RedPixel Engine");
         let window: Box<dyn Window> = event_loop.create_window(attributes)?;
-        self.window = Some(window);
-        Ok(())
+
+        Ok(Self {
+            window: Arc::from(window),
+        })
+    }
+
+    pub fn get_window(&self) -> Arc<dyn Window> {
+        self.window.clone() // TODO: No need for cloning.
+    }
+
+    pub fn handle_window_event(&self, event: &WindowEvent) {
+        match event {
+            WindowEvent::Focused(..) => {}
+            WindowEvent::ScaleFactorChanged { .. } => {}
+            _ => {}
+        }
     }
 
     pub fn request_redraw(&self) {
-        if let Some(window) = &self.window {
-            window.request_redraw();
-        }
+        self.window.request_redraw();
     }
 
-    pub fn event_handler(&mut self, event_loop: &dyn ActiveEventLoop, event: &WindowEvent) {
-        match event {
-            WindowEvent::CloseRequested | WindowEvent::Destroyed => event_loop.exit(),
-            WindowEvent::Focused(_is_focused) => {}
-            WindowEvent::SurfaceResized(_size) => {}
-            _ => {}
-        }
+    pub fn is_window_event(&self, event: &WindowEvent) -> bool {
+        matches!(event, WindowEvent::Focused { .. } | WindowEvent::ScaleFactorChanged { .. })
     }
 }